% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuate.R
\name{attenuate}
\alias{attenuate}
\title{Calculate light attenuation in water}
\usage{
attenuate(PAR, z, prop.reflect, K, Kp, Bz, selfshading)
}
\arguments{
\item{PAR}{photosynthetically active radiation at the water surface, in microEinsteins.}

\item{z}{water depth in meters.}

\item{prop.reflect}{proportion of light reflected by the water surface (default 0.1).}

\item{K}{light attenuation coefficient of water (default 0.12)}

\item{Kp}{light attenuation coefficient of plant material.  GenVeg uses a default of 0.02 m2/g, but they use the following for crops: rice 0.38 m2/g, wheat 0.48 m2/g, corn 0.51 m2/g, citing Chakraborty et al. 2018. GenVeg's NetLogo photosynthesis function uses 0.0235 citing van Nes et al. 2003 Eco Mod which provided that value for Vallisneria.}

\item{Bz}{plant biomass above depth z}

\item{selfshading}{a logical value indicating whether the macrophyte experiences self-shading (TRUE) or not (FALSE).}
}
\value{
a vector of three PAR values in microEinsteins, corresponding to light available to macrophyte leaves at three representive times of day, roughly midday, morning/afternoon, and dawn/dusk.
}
\description{
\code{attenuate} calculates available photosynthetically active radiation (PAR) at the water surface, taking into account reflection off the water surface and attenuation with depth.
}
\details{
Given photosynthetically active radiation (PAR) available at the water surface, attenuate calculates PAR at any water depth, including losses of PAR due to reflection off the water surface, attenuation in the water column, and optional self-shading by macrophytes.  Default values are taken from GenVeg (NetLogo version, photosynthesis function), who cite (http://www.lakeaccess.org/ecology/lakeecologyprim3.html) for the light attenuation coefficient K, and (http://www.esf.edu/efb/schulz/Limnology/Light.html) for the equation without self-shading.
}
\examples{
#Result: Vector of c(2394.6852, 1596.4568, 399.1142) microEinsteins of PAR
attenuate(PAR=c(3000,2000,500), z=1, prop.reflect=0.1, K=0.12, selfshading=FALSE)

#Result: Vector of c(2255.2296, 1503.4864, 375.8716) microEinsteins of PAR
attenuate(PAR=c(3000,2000,500), z=1, prop.reflect=0.1, K=0.12, Kp=0.2, Bz=0.3, selfshading=TRUE)


}
\references{
Chakraborty, P. K.; Banerjee, S.; Mukherjee, A.; Nath, R.; and Samanta, S. 2018. Extinction coefficient and photosynthetically active radiation use efficiency of summer rice as influenced by transplanting dates. Journal of Environmental Biology 39: 467-471. DOI: http://doi.org/10.22438/jeb/39/4/MRN-661.
van Nes, Egbert H; Scheffer, Marten; van den Berg, Marcel S., and Coops, Hugo. 2003. Charisma: a spatial explicit simulation model of submerged macrophytes. Ecological Modelling 159: 103-116.
http://www.lakeaccess.org/ecology/lakeecologyprim3.html
http://www.esf.edu/efb/schulz/Limnology/Light.html
}
