% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakage.R
\name{breakage}
\alias{breakage}
\title{Break stems of unrealistic stem length}
\usage{
breakage(stem.cm, max.stem.cm)
}
\arguments{
\item{stem.cm}{stem length in cm.}

\item{max.stem.cm}{physiological maximum of stem length in cm.}
}
\value{
daily macrophyte loss due to breakage, a positive number.
}
\description{
\code{breakage} removes stem length for physiologically unrealistic stem length
}
\details{
This function removes stem length in excess of a user-defined physiological maximum, returning
unrealistically high stem lengths to their maximum allowed value. Parameterization is based on
observations that most Podostemum stems are less than 20 cm long (Hammond 1937), and we rarely
observe stems longer than ~30 cm. This function could also be applied to biomass, in which case a
reasonable upper limit for a single stem is ~0.4 g ash-free dry mass.

#' @references
Hammond, Bayard L. 1937. Development of Podostemon ceratophyllum. Bulletin of the Torrey Botanical
Club 64(1): 17-36.
}
\examples{
#Result: 20 cm
breakage(20, 50)

#Result: 50 cm
breakage(70, 50)

#Result: vector of values with max of 50 cm
breakage(seq(10,100,10), 50)

}
