% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{daylength}
\alias{daylength}
\title{Calculate day length}
\usage{
daylength(day, lat)
}
\arguments{
\item{day}{Julian day}

\item{lat}{latitude in decimal degrees}
}
\value{
daylength, the number of hours between sunrise and sunset.
}
\description{
\code{daylength} calculates day length in hours based on latitude and Julian day
}
\details{
Given Julian day and latitude, this function calculates day length in hours following GenVeg (CITE, plus their sources).
}
\examples{
#Result: 12.90703
daylength(120,25)

#Result: 11.74682
daylength(200,-5)

}
\references{
GenVeg
Astro?
}
