% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbivory.R
\name{herbivory}
\alias{herbivory}
\title{Calculate macrophyte losses to herbivory based on empirical data}
\usage{
herbivory(old.size, H, Vlow, Vhigh, V, type)
}
\arguments{
\item{old.size}{size at prior time step.}

\item{H}{maximum herbivory rate in proportion lost per day.}

\item{Vlow}{flow velocity (in m/s) below which herbivory is at the maximum rate.}

\item{Vhigh}{flow velocity (in m/s) above which herbivory is 0.}

\item{V}{flow velocity in m/s.}

\item{type}{an integer between 1 and 3 specifying the numerical form of the herbivory model.}
}
\value{
size of macrophyte after herbivory
}
\description{
\code{herbivory} calculates the size of a macrophyte after losses due to herbivory.
}
\details{
This function calculates macrophyte size (stem length or biomass) following losses due to herbivory in a
given time step. The extent of herbivory is considered to be negatively related to flow velocity (Wood
et al. 2019). This function allows the velocity-herbivory relationship to take multiple forms, described
below, where old.size is the size of the macrophyte at the beginning of the time step, H is the maximum
herbivory rate in proportion lost per day, Vlow is the flow velocity (in m/s) below which herbivory
is at its maximum rate H, Vhigh is the flow velocity (in m/s) above which herbivory ceases, and V
is the flow velocity in the current time step, in m/s. All measures of macrophyte size must use the same
units.

Type 1: Constant herbivory.

Type 2: Herbivory is driven by a threshold in velocity.

Type 3: Herbivory is described by a step function, starting at its maximum rate for velocities less
than Vlow, and declining linearly to reach zero when velocity reaches Vhigh.
}
\examples{
#Result: 2
herbivory(20, H=0.1, Vlow=0.5, Vhigh=2.5, V=1, type=1)

#Result: 0.1
herbivory(0.2, H=0.5, Vlow=0.5, Vhigh=2.5, V=1.0, type=2)

#Result: 0
herbivory(0.2, H=0.5, Vlow=0.5, Vhigh=2.5, V=2.7, type=2)

#Result: 6
herbivory(20, H=0.3, Vlow=0.5, Vhigh=2.5, V=0.2, type=3)

#Result: 4.2
herbivory(20, H=0.3, Vlow=0.5, Vhigh=2.5, V=1.1, type=3)

#Result: 0
herbivory(20, H=0.3, Vlow=0.5, Vhigh=2.5, V=2.6, type=3)

}
\references{
Wood JL, Skaggs JW, Conn C, Freeman MC. 2019. Water velocity regulates macro-consumer herbivory on
the benthic macrophyte Podostemum ceratophyllum Michx. Freshwater Biology 64(11): 2037-2045.
DOI: 10.1111/fwb.13393.
}
