% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{light}
\alias{light}
\title{Calculate photosynthetically active radiation}
\usage{
light(day, lat)
}
\arguments{
\item{day}{Julian day}

\item{lat}{latitude in decimal degrees}
}
\value{
tmpvec a vector of three PAR values in microEinsteins
}
\description{
\code{light} calculates photosynthetically active radiation (PAR) at the water surface
at three times of day based on latitude and Julian day
}
\details{
Given Julian day and latitude, this function calculates photosynthetically active radiation (PAR) in microEinsteins at the water surface following GenVeg (CITE, plus their sources).  This function returns a vector of PAR values corresponding to three times of day, roughly midday, morning/afternoon, and dawn/dusk.
}
\examples{
#Result: Stem length of 20.63898 centimeters
biomass.to.stemlength.pod(0.1, type=1)

#Result: Stem length of 43.11088 centimeters
biomass.to.stemlength.pod(0.1, type=2)

#Result: Stem length of 14.82514 centimeters
biomass.to.stemlength.pod(0.1, type=3)

#Result: Error message indicating incorrect model specification
biomass.to.stemlength.pod(0.1, type=7)

#Result: Warning message indicating unrealistic biomass
#biomass.to.stemlength.pod(1, type=1)

}
\references{
GenVeg
Astro?
}
