% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{Calculate macrophyte mortality}
\usage{
mortality(old.size, mort.rate)
}
\arguments{
\item{old.size}{initial biomass of macrophyte}

\item{mort.rate}{mortality rate of macrophyte, in g/g/day}
}
\value{
daily macrophyte biomass loss due to mortality, a positive number.
}
\description{
\code{mortality} calculates macrophyte biomass loss due to mortality
}
\details{
Calculates plant mortality, following NetLogo version of mortality rather than R GenVeg version. Note GenVeg uses mortality rate of 0.021 /d (that is, 2.1\% loss per day), and the GenVeg NetLogo mortality file contains some interesting simple respiration functions.
}
\examples{
#Result: Stem length of 20.63898 centimeters
biomass.to.stemlength.pod(0.1, type=1)

#Result: Stem length of 43.11088 centimeters
biomass.to.stemlength.pod(0.1, type=2)

#Result: Stem length of 14.82514 centimeters
biomass.to.stemlength.pod(0.1, type=3)

#Result: Error message indicating incorrect model specification
biomass.to.stemlength.pod(0.1, type=7)

#Result: Warning message indicating unrealistic biomass
#biomass.to.stemlength.pod(1, type=1)

}
\references{
Rack, Laura. 2022. River Basin Center, University of Georgia. Unpublished data.
}
