% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis.R
\name{photosynthesis}
\alias{photosynthesis}
\title{Calculate growth due to photosynthesis}
\usage{
photosynthesis(
  PAR,
  pbiomass,
  Pmax,
  daylength,
  glucosereq,
  Hi = 0,
  temp,
  D = 1,
  Hd = 0,
  C = 1,
  Hc = 0,
  N = 1,
  Hn = 0
)
}
\arguments{
\item{PAR}{a vector of three values of photosynthetically active radiation reaching macrophyte leaves at three representative times of day, in microEinsteins.}

\item{pbiomass}{biomass of photosynthetic tissue, in g}

\item{Pmax}{maximum photosynthetic rate in g glucose / g photosynthetic tissue / day. This quantity is species specific; GenVeg provides the following values: rice 0.027, wheat 0.5, corn 0.0372.}

\item{daylength}{length of day in hours from sunrise to sunset.}

\item{glucosereq}{glucose requirement for growth. GenVeg, citing Teh (2006) Table 7.4, give values of 1.436 for leaves, 1.513 for stems, and 1.444 for roots.}

\item{Hi}{half-saturation constant for light, in units of PAR (microEinsteins).  GenVeg (NetLogo) photosynthesis function suggests 14 for C. aspera and 52 for P. pectinatus.}

\item{temp}{average daily temperature in °C.}

\item{D}{depth of photosynthetic tissue below the water surface (?), in same units as Hd.}

\item{Hd}{half-saturation constant for depth, in same units as D. GenVeg (NetLogo) photosynthesis function describes this as the half-saturation constant for light diffusion and uses a default of 1 m.}

\item{C}{ambient carbonate concentration, in same units as Hc.}

\item{Hc}{half-saturation constant for carbonate, in same units as C. GenVeg (NetLogo) photosynthesis function suggests 60, but notes that it varies by species.}

\item{N}{ambient concentration of limiting nutrient, in same units as Hn.}

\item{Hn}{half-saturation concentration of limiting nutrient, in same units as N.}
}
\value{
daily biomass gain due to gross photosynthetic assimilation.
}
\description{
\code{photosynthesis} calculates the amount of biomass gained by a macrophyte on a given day, given a species-specific maximum photosynthetic rate and optional parameters to account for growth limitation by temperature, light, depth, carbonate, and nutrients.
}
\details{
This function calculates gross photosynthetic assimilation in units of macrophyte biomass given light availability, plant height, river depth, and several plant parameters.  It is designed for use with the light, daylength, and attenuate functions to calculate light availability for photosynthesis.  It is designed to be used alongside the respiration function (which calculates maintenance respiration) to calculate net photosynthetic biomass assimilation.  This function assumes Michaelis-Menten or Monod dynamics for light, carbonate, and nutrient assimilation, as well as depth, following the Charisma model (van Nes et al. 2003). Temperature is assumed to limit photosynthesis by a Hill function.  This function applies to a single plant organ or tissue, but can be looped over different organs or tissues with different photosynthetic properties.
}
\examples{
#Result: biomass gain of 0.9980636 g
photosynthesis(PAR=c(3000, 2000, 500), pbiomass=1, Pmax=0.2, daylength=13, glucosereq=1.5, Hi=30, temp=17)

}
\references{
GenVeg

Carr, Geneviève M.; Duthie, Hamish C.; and Taylor, William D. 1997. Models of aquatic plant productivity: a review of the factors that influence growth. Aquatic Botany 59: 195-215.
Chakraborty, P. K.; Banerjee, S.; Mukherjee, A.; Nath, R.; and Samanta, S. 2018. Extinction coefficient and photosynthetically active radiation use efficiency of summer rice as influenced by transplanting dates. Journal of Environmental Biology 39: 467-471. DOI: http://doi.org/10.22438/jeb/39/4/MRN-661.
Teh, Christopher B. S. 2006. Introduction to mathematical modeling of crop growth: How the equations are derived and assembled into a computer model. Boca Raton, FL: BrownWalker Press.
van Nes, Egbert H; Scheffer, Marten; van den Berg, Marcel S., and Coops, Hugo. 2003. Charisma: a spatial explicit simulation model of submerged macrophytes. Ecological Modelling 159: 103-116.
}
