% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{daylength}
\alias{daylength}
\title{Calculate day length}
\usage{
daylength(day, lat)
}
\arguments{
\item{day}{Julian day}

\item{lat}{latitude in decimal degrees}
}
\value{
daylength, the number of hours between sunrise and sunset.
}
\description{
\code{daylength} calculates day length in hours based on latitude and Julian day
}
\details{
Given Julian day and latitude, this function calculates day length in hours following GenVeg (CITE, plus their sources).
}
\examples{
#Result: day length of 12.90703 hours
daylength(120, 25)

#Result: day length of 12.28713 hours
daylength(180, 5)

#Result: day length of 5.837386 hours
daylength(5, 60)

}
\references{
GenVeg
Astro?
}
