% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{daylength}
\alias{daylength}
\title{Calculate day length}
\usage{
daylength(day, lat)
}
\arguments{
\item{day}{Julian day}

\item{lat}{latitude in decimal degrees}
}
\value{
daylength, the number of hours between sunrise and sunset.
}
\description{
\code{daylength} calculates day length in hours based on latitude and Julian day
}
\details{
Given Julian day and latitude, this function calculates day length in hours following GenVeg (CITE, plus their sources).
}
\examples{
#Result: Stem length of 20.63898 centimeters
biomass.to.stemlength.pod(0.1, type=1)

#Result: Stem length of 43.11088 centimeters
biomass.to.stemlength.pod(0.1, type=2)

#Result: Stem length of 14.82514 centimeters
biomass.to.stemlength.pod(0.1, type=3)

#Result: Error message indicating incorrect model specification
biomass.to.stemlength.pod(0.1, type=7)

#Result: Warning message indicating unrealistic biomass
#biomass.to.stemlength.pod(1, type=1)

}
\references{
GenVeg
Astro?
}
