% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{light}
\alias{light}
\title{Calculate photosynthetically active radiation}
\usage{
light(day, lat)
}
\arguments{
\item{day}{Julian day}

\item{lat}{latitude in decimal degrees}
}
\value{
tmpvec a vector of three PAR values in microEinsteins
}
\description{
\code{light} calculates photosynthetically active radiation (PAR) at the water surface
at three times of day based on latitude and Julian day
}
\details{
Given Julian day and latitude, this function calculates photosynthetically active radiation (PAR) in microEinsteins at the water surface following GenVeg (CITE, plus their sources).  This function returns a vector of PAR values corresponding to three times of day, roughly midday, morning/afternoon, and dawn/dusk.
}
\examples{
#Result: a vector of the following three PAR values in microEinsteins: 
#2879.4155, 1882.2661, 382.6393
light(120,25)

}
\references{
GenVeg
Astro?
}
