% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{Calculate macrophyte mortality}
\usage{
mortality(old.size, mort.rate)
}
\arguments{
\item{old.size}{initial biomass of macrophyte}

\item{mort.rate}{mortality rate of macrophyte, in g/g/day}
}
\value{
biomass of macrophyte after mortality
}
\description{
\code{mortality} calculates macrophyte biomass loss due to mortality
}
\details{
Calculates plant mortality, following NetLogo version of mortality rather than R GenVeg version. Note GenVeg uses mortality rate of 0.021 /d (that is, 2.1\% loss per day), and the GenVeg NetLogo mortality file contains some interesting simple respiration functions.
}
\examples{
#Result: Biomass of 1.96 g
mortality(old.size=2, mort.rate=0.02)

}
\references{
Rack, Laura. 2022. River Basin Center, University of Georgia. Unpublished data.
}
