% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiration.R
\name{respiration}
\alias{respiration}
\title{Calculate maintenance respiration}
\usage{
respiration(kmprime, temp, liveweight, totalweight, glucosereq)
}
\arguments{
\item{kmprime}{base maintenance respiration coefficient at 25 °C. GenVeg, citing Teh (2006) table 7.1, give a value of 0.03 for leaves and 0.015 for stems and roots.  GenVeg's NetLogo code calls this figure q10 instead, treats stems and leaves together, and uses the average value of 0.0225.}

\item{temp}{average daily temperature in °C.}

\item{liveweight}{live biomass of macrophyte or its relevant organ or tissue.}

\item{totalweight}{total (live plus dead) biomass of macrophyte or its relevant organ or tissue.}

\item{glucosereq}{glucose requirement for growth. GenVeg, citing Teh (2006) Table 7.4, give values of 1.436 for leaves, 1.513 for stems, and 1.444 for roots.}
}
\value{
daily biomass loss due to maintenance respiration, a positive number.
}
\description{
\code{respiration} calculates the amount of biomass lost by a macrophyte on a given day due to maintenance respiration.
}
\details{
This function calculates macrophyte biomass losses due to maintenance respiration, and is intended to be used alongside the photosynthesis function to convert gross photosynthetic assimilation to net photosynthetic assimilation.  This function applies to a single plant organ or tissue, but can be looped over different organs or tissues with different metabolic properties.  Positive values of respiration represent biomass lost to respiration, so net photosynthesis = gross photosynthesis - respiration.
Inclusion of standing dead biomass via the totalweight argument is carried over from GenVeg, where standing dead biomass was used as a proxy of plant age, and maintenance respiration assumed to decrease in older plants.  The effects of standing dead biomass in other circumstances should be considered with caution.
}
\examples{
#Result: macrophyte biomass loss of 0.015 g
respiration(kmprime=0.0225, temp=15, liveweight=2, totalweight=2, glucosereq=1.5)

#Result: macrophyte biomass loss of 0.01 g
respiration(kmprime=0.0225, temp=15, liveweight=2, totalweight=3, glucosereq=1.5)

#Result: macrophyte biomass loss of 0.02222222 g
respiration(kmprime=0.05, temp=15, liveweight=2, totalweight=3, glucosereq=1.5)

}
\references{
GenVeg
Teh, Christopher B. S. 2006. Introduction to mathematical modeling of crop growth: How the equations are derived and assembled into a computer model. Boca Raton, FL: BrownWalker Press.
}
